requirements.txt
python-dotenv==1.0.1
pymongo==4.8.0
aioconsole==0.8.0